Large numbers
This commit provides another definition of large numbers, that has
better properties than the definition from `UnaryFixpoint`. We found
the notion of large numbers useful, so it becomes part of the public
interface.

The old definition of large numbers is still needed, as the new
definition is based on guarded recursion, which requires the old
definition.